Hacked mapc to avoid i18n.
We can drop INTL_LIBS from BASE_LIBS in the Makefile, but then we also
have to make sure that all the  included headers are clean of any i18n
magic.

I call  it a "hack"  because this should really  be sorted out  in the
Makefile, not source code. Unfortunately it seems to be the "right but
complicated" way.

Removed a similar hack from democonv.  No longer needed.


git-svn-id: https://s.snth.net/svn/neverball/trunk@1284 78b8d119-cf0a-0410-b17c-f493084dd1d7