Only save set score when player clicks "Finish" in Goal screen.
The problem with saving the set score during progress_stat (called at
the end of each level) was that it was saved multiple times if the
player chose to replay the last level of the set.

For this purpose, progress_exit was redefined to match its name and
really "exit progress" (which includes saving set score) instead of
being just a short-hand for "progress_stat(); progress_stop()".


git-svn-id: https://s.snth.net/svn/neverball/trunk@1999 78b8d119-cf0a-0410-b17c-f493084dd1d7