Use SDL_GL_SWAP_CONTROL for vertical sync
Since 1.2.10 (released May 17, 2006), SDL provides a SWAP_CONTROL GL
attribute for controlling the buffer swap interval.  This patch removes
most of the platform-specific code in share/sync and uses the attribute
instead.  (Code for Mac OS X still remains, though unused, because it's
not clear if SWAP_CONTROL is implemented on OS X.)

git-svn-id: https://s.snth.net/svn/neverball/trunk@2417 78b8d119-cf0a-0410-b17c-f493084dd1d7