Correct logic of BSP back/front tests
This patch corrects faulty logic in the BSP "in front of plane" and
"behind plane" tests.  Previously, if the ball was determined to not
be on the currently-tested side of the dividing plane, instead of
failing, the test also checked if the ball was moving towards the
plane, and succeeded in that case.  Instead, the code now tests the
position of the ball before and after movement and passes the tests if
the ball either is on the corresponding side, or isn't, but will be
after DT seconds.

git-svn-id: https://s.snth.net/svn/neverball/trunk@2512 78b8d119-cf0a-0410-b17c-f493084dd1d7