po/zh_CN: translate font path to "ttf/wqy-zenhei.ttc"
The default DejaVu font appears to be utterly useless for Chinese.
The translator suggests to use the "WenQuanYi Zen Hei" font instead.
Unfortunately, the font file (as found in Debian's ttf-wqy-zenhei
package) is approximately 13 MB in size, so distributing it with the
game doesn't seem a very appealing idea.

This patch goes half-way and merely advertises the correct font
(causing an error message to be printed).  Packagers and users wishing
to support Simplified Chinese should install the font to
"data/ttf/wqy-zenhei.ttc".

git-svn-id: https://s.snth.net/svn/neverball/trunk@2846 78b8d119-cf0a-0410-b17c-f493084dd1d7