Pass the interpolation blend factor differently
Rather than pass it to the draw function, make it a member of the lerp
structure and set it through a function call.  Doing it this way
eliminates a few ill defined cases: 1) transitions between screens
that use different blend factors; 2) screens that render game state
from either a replay or a simulation.

git-svn-id: https://s.snth.net/svn/neverball/trunk@3490 78b8d119-cf0a-0410-b17c-f493084dd1d7