Make the darwin build a tiny bit smarter.
Xcode 5 pushes LLVM and diverges a bit from the behavior of prior
versions. It no longer looks in /usr/local/lib by default. This
change to the makefile accommodates the new behavior while still
defaulting to the use of macports in /opt. It also ignores either
alternative if not present. This is trivially extensible to fink
et al.