Remove billboard flag 4 (additive blending)
All it does is set a material flag that overrides the material file,
causing confusion when somebody tries to add/remove the additive flag
from the material file.

Due to a flag being removed, B_NOFACE, which was 8, became 4. Normally I
would worry about binary compatibility and keep it an 8, but chances are
that both these flags have never been used outside of the official map
files.